refactor(core): Improve DP_Optimizations.H and examples
Refactor DP optimization and tree DP headers, enhancing readability and consistency.
Update examples for knapsack, LIS, matrix chain, subset sum, and tree DP to align with new structures and use more modern C++ practices (e.g., range-based for loops, explicit loop bounds).
Address C++ casting best practices with clamped_cast.
Normalize main function signatures across examples.

+++ Tests/subset_sum_test.cc
@@ -69,7 +69,7 @@ namespace
   {
     int sum = 0;
     BitArray seen(vals.size());
-    for (unsigned long idx : selected)
+    for (size_t idx : selected)
       {
         if (idx >= vals.size() or seen[idx])
           return false;
@@ -108,7 +108,7 @@ TEST(SubsetSum, BasicFound)
 
   // Verify selected indices sum to target
   int total = 0;
-  for (unsigned long selected_indice : r.selected_indices)
+  for (size_t selected_indice : r.selected_indices)
     total += vals[selected_indice];
   EXPECT_EQ(total, 9);
 }
